modify grpc types (#113)
* rm some unused and add missed struct

* add point in some struct

* update grpcTypes

* add ShardStatus struct

* AddBlockListForSyncResponse.MinorBlockHashList

* add Branch in AddBlockListForSyncRequest

* change ShardStatus.LastTime from uint32 to uint64

+++ core/minorblockchain_addon.go
@@ -1049,7 +1049,7 @@ func (m *MinorBlockChain) GetShardStatus() (*rpc.ShardStatus, error) {
 	txCount := uint32(0)
 	blockCount := uint32(0)
 	staleBlockCount := uint32(0)
-	lastBlockTime := uint32(0)
+	lastBlockTime := uint64(0)
 	for cblock.IHeader().NumberU64() > 0 && cblock.IHeader().GetTime() > cutoff {
 		txCount += uint32(len(cblock.GetTransactions()))
 		blockCount++
@@ -1059,7 +1059,7 @@ func (m *MinorBlockChain) GetShardStatus() (*rpc.ShardStatus, error) {
 			return nil, ErrMinorBlockIsNil
 		}
 		if lastBlockTime == 0 {
-			lastBlockTime = uint32(m.CurrentBlock().IHeader().GetTime() - cblock.IHeader().GetTime())
+			lastBlockTime = m.CurrentBlock().IHeader().GetTime() - cblock.IHeader().GetTime()
 		}
 	}
 	if staleBlockCount < 0 {
